develop/payment_refactor - feat : CartDisassembler 단위 테스트

+++ src/main/java/com/daniel/mychickenbreastshop/usecase/orderpayment/application/strategy/service/kakaopay/KakaopayStrategyApplication.java
@@ -76,7 +76,7 @@ public class KakaopayStrategyApplication implements PaymentStrategyApplication<P
                 savedProduct.getImage(),
                 savedProduct.getContent()
         );
-        Payment payment = Payment.createPayment((long) itemPayRequestDto.getTotalAmount());
+        Payment payment = Payment.createPayment(itemPayRequestDto.getTotalAmount());
 
         order.addOrderProduct(orderProduct);
         order.setPaymentInfo(payment);
